Move navigateDashboard action to navigateActions
